further navigation and password hasher

+++ src/com/Connector/DatabaseConnector.java
@@ -14,30 +14,19 @@ public class DatabaseConnector {
 
     // -------- CRUD User -------- 
     public void addUser(String username, String password){
-        if(username.length()> 32){
-            System.out.println("The chosen username is too long. It must be a maximum of 32 characters.");
-            return;
-        }
         dbConn.executeStatement("select Name from User where Name = '" + username + "'");
         QueryResult qr = dbConn.getCurrentQueryResult();
         if(qr.getRowCount() != 0){
             System.out.println("The chosen username is already taken. Please choose another!");
             return;
-        }
-        if(password.length()> 64){
-            System.out.println("The chosen password is too long. It must be a maximum of 64 characters.");
-            return;
-        }
-        
+        }        
         dbConn.executeStatement("Insert into User (Name, Password) Values ('"+ username +"','"+ password +"')");
-        System.out.println("The user with username: '" + username + "' was successfully created.");
     }
 
     public void deleteUser(String username){
         int id = findUserByName(username);
         if(id==0) return; // --> User not found
         dbConn.executeStatement("delete from User where id = " + id);
-        System.out.println("The user with username: '" + username + "' was successfully deleted.");
     }
 
     //returns ID of a user
@@ -47,6 +36,7 @@ public class DatabaseConnector {
         return Integer.parseInt(qr.getData()[0][0]);
     }
     
+    // returns ID of a user
     public int findUserByID(int ID){
         dbConn.executeStatement("select ID from User where ID = " + ID);
         QueryResult qr = dbConn.getCurrentQueryResult();
@@ -57,6 +47,7 @@ public class DatabaseConnector {
         return Integer.parseInt(qr.getData()[0][0]);
     }
 
+    // return password of a user
     public String getPassword(int id){
         String result = "";
         dbConn.executeStatement("select Password from User where id = "+ findUserByID(id));
@@ -67,31 +58,42 @@ public class DatabaseConnector {
         return result = qr.getData()[0][0];
     }
 
+    public String[] getUser(){
+        dbConn.executeStatement("select Name from User");
+        QueryResult qr = dbConn.getCurrentQueryResult();
+        int rowCount = qr.getRowCount();
+        String[] result = new String[rowCount];
+        for(int i = 0; i < rowCount; i++){
+            result[i] = qr.getData()[i][0];
+        }
+        return result; 
+    }
+
+
+
 
 
     // -------- CRUD Category -------- 
     public void createCategory(String categoryName){
-        if(categoryName.length() > 32){
-            System.out.println("The chosen category name is too long. It must be a maximum of 32 characters.");
-            return;
-        }
         dbConn.executeStatement("Insert into Category (Description) Values ('"+ categoryName +"')");
         System.out.println("The category with name: '" + categoryName + "' was successfully created.");
     }
 
     public void deleteCategory(String description){
         int id = findCategoryByDescription(description);
-        if(id==0) return; // --> Category not found
+        if(id==0 || id == 1) return; // --> Category not found
         dbConn.executeStatement("delete from Category where id = " + id);
         System.out.println("The category with name: '" + description + "' was successfully deleted.");
     }
 
+    // returns ID of a category
     public int findCategoryByDescription(String description){
         dbConn.executeStatement("select ID from Category where Description = '"+ description +"'");
         QueryResult qr = dbConn.getCurrentQueryResult();
         return Integer.parseInt(qr.getData()[0][0]);
     }
 
+    // returns ID of a category
     public int findCategoryByID(int ID){
         dbConn.executeStatement("select ID from Category where ID = '"+ ID +"'");
         QueryResult qr = dbConn.getCurrentQueryResult();
@@ -101,6 +103,22 @@ public class DatabaseConnector {
         } 
         return Integer.parseInt(qr.getData()[0][0]);
     }
+
+    // returns an array with all categories
+    public String[] getCategories(){
+        dbConn.executeStatement("select Description from Category");
+        QueryResult qr = dbConn.getCurrentQueryResult();
+        int rowCount = qr.getRowCount();
+        String[] result = new String[rowCount];
+        for(int i = 0; i < rowCount; i++){
+            result[i] = qr.getData()[i][0];
+        }
+        return result; 
+    }
+
+
+
+
 
 
     // -------- CRUD ToDo -------- 
@@ -122,6 +140,7 @@ public class DatabaseConnector {
             }
             dbConn.executeStatement("Insert into ToDo (Description, Title, Date, Important, CategoryID, OwnerID)"+
                                     "Values ('" + description + "','" + title +"',NOW()," + important + "," + categoryID + "," + ownerID +")");
+                                    //'description'
             setSharedConnection(ownerID, findToDoByTitle(title));
             System.out.println("The ToDo '" + title + "' was successfully created.");
         } else {
@@ -130,16 +149,17 @@ public class DatabaseConnector {
         }
     }
 
-    public void deleteToDo(int id){
-        if(findToDoByID(id) == 0) return; // ToDo not found
-        dbConn.executeStatement("select OwnerID from ToDo where ID = " + id);
+    public void deleteToDo(int toDoID){
+        if(findToDoByID(toDoID) == 0) return; // ToDo not found
+        dbConn.executeStatement("select OwnerID from ToDo where ID = " + toDoID);
         QueryResult qr = dbConn.getCurrentQueryResult();
         int ownerID = Integer.parseInt(qr.getData()[0][0]);
-        deleteSharedConnection(ownerID, id);
-        dbConn.executeStatement("delete from ToDo where id = " + id);
-        System.out.println("The ToDo with ID: '" + id + "' was successfully deleted.");
+        deleteSharedConnection(ownerID, toDoID);
+        dbConn.executeStatement("delete from ToDo where id = " + toDoID);
+        System.out.println("The ToDo with ID: '" + toDoID + "' was successfully deleted.");
     }
 
+    // returns ID of a todo
     public int findToDoByTitle(String title){
         dbConn.executeStatement("select ID from ToDo where title = '" + title + "'");
         QueryResult qr = dbConn.getCurrentQueryResult();
@@ -150,19 +170,34 @@ public class DatabaseConnector {
         return Integer.parseInt(qr.getData()[0][0]);
     }
 
-    public int findToDoByID(int id){
-        dbConn.executeStatement("select ID from ToDo where id = " + id);
+    // returns ID of a todo
+    public int findToDoByID(int toDoID){
+        dbConn.executeStatement("select ID from ToDo where id = " + toDoID);
         QueryResult qr = dbConn.getCurrentQueryResult();
         if(qr.getRowCount() == 0){
-            System.out.println("The ToDo with ID: '" + id + "' was not found.");
+            System.out.println("The ToDo with ID: '" + toDoID + "' was not found.");
             return 0;
         }
         return Integer.parseInt(qr.getData()[0][0]);
     }
 
-    public void printToDo(int id){
-        
+    // reteurns an array with all todos of transferred ownerID
+    public String[] getToDos(int ownerID){
+        dbConn.executeStatement("select title "+
+                                "from ToDo "+
+                                "where OwnerID = " + ownerID +
+                                " order by Date asc");
+        QueryResult qr = dbConn.getCurrentQueryResult();
+        int rowCount = qr.getRowCount();
+        String[] result = new String[rowCount];
+        for(int i = 0; i < rowCount; i++){
+            result[i] = qr.getData()[i][0];
+        }
+        return result;
     }
+
+
+
 
 
 
@@ -181,5 +216,17 @@ public class DatabaseConnector {
         if(findUserByID(userID)==0) return;
         dbConn.executeStatement("DELETE FROM Shared_ToDo_Users WHERE user_id = "+userID+" AND todo_id = "+toDoID);
         System.out.println("The user with ID: '" + userID + "' has nolonger access to the ToDo with ID: " + toDoID + "'");
+    }
+
+    // returns an array with all todoIDs of the transferred ownerID
+    public int[] getSharedConnections(int ownerID){
+        dbConn.executeStatement("select todo_id from Shared_ToDo_Users where user_id = "+ownerID);
+        QueryResult qr = dbConn.getCurrentQueryResult();
+        int rowCount = qr.getRowCount();
+        int[] result = new int[rowCount];
+        for(int i = 0; i < rowCount; i++){
+            result[i] = Integer.parseInt(qr.getData()[i][0]);
+        }
+        return result;
     }
 }
